# docs/platform/taskr/taskr-openapi.yaml
openapi: 3.1.0
info:
  title: TaskR API
  version: 0.1.0
  description: |
    Draft OpenAPI specification for TaskR endpoints. Generated from docs/platform/taskr/api-reference.md.
servers:
  - url: https://{environment}.taskr.dydact.io/api
    variables:
      environment:
        default: staging
        enum: [dev, staging, prod]
  - url: https://{environment}.scr.dydact.io/scr/api
    description: scrAIv bridge endpoints
    variables:
      environment:
        default: staging
        enum: [dev, staging, prod]
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantId:
      name: x-tenant-id
      in: header
      required: true
      schema:
        type: string
    UserId:
      name: x-user-id
      in: header
      required: false
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            detail:
              type: object
          required: [code, message]
...
    Task:
      type: object
      properties:
        id:
          type: string
        list_id:
          type: string
        title:
          type: string
        status:
          type: string
        assignees:
          type: array
          items:
            type: string
        due_date:
          type: string
          format: date
        priority:
          type: string
        custom_fields:
          type: array
          items:
            type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, list_id, title, status, created_at, updated_at]
...
    TaskRequest:
      type: object
      properties:
        list_id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        assignees:
          type: array
          items:
            type: string
        due_date:
          type: string
          format: date
        priority:
          type: string
        custom_fields:
          type: array
          items:
            type: object
      required: [list_id, title]
paths:
...
  /tasks/{id}:
    get:
      summary: Get task detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update task
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/UserId'
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...